requirements.txt
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
findspark==2.0.1
fonttools==4.40.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
py4j==0.10.9.7
pymongo==4.3.3
pyparsing==3.0.9
pyspark==3.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
tzdata==2023.3  